requirements.txt
flask
flask-cors
flask-sqlalchemy
pymysql
python-dotenv
openai
requests
PyPDF2
